requirements.txt
selenium
selenium-stealth
webdriver-manager
google-genai
python-dotenv
feedparser
schedule